# Conflicts:
#	backend/src/main/java/com/smalltrend/controller/pos/ProductVariantController.java
#	backend/src/main/java/com/smalltrend/dto/Module1/ProductVariantRespone.java
#	backend/src/main/java/com/smalltrend/service/Module1/ProductVariantService.java
#	frontend/src/App.jsx
#	frontend/src/components/layout/Sidebar.jsx

+++ backend/src/main/java/com/smalltrend/repository/ProductRepository.java
@@ -2,8 +2,11 @@ package com.smalltrend.repository;
 
 import com.smalltrend.entity.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
+
 @Repository
-public interface ProductRepository extends JpaRepository<Product, Long> {
+public interface ProductRepository extends JpaRepository<Product, Integer> {
 }
